修正 overlay FPS：使用影片原生 FPS 取代固定 30
原本 overlay 永遠以 DEFAULT_FPS=30 渲染，非 30fps 影片的 overlay
時長會與裁切片段不一致（如 60fps 影片 2 秒片段 → overlay 變 4 秒）。
修正：上傳時探測 native_fps 存入 TaskState，一路傳到 render_overlay_video
與前端 detection_fps，確保 overlay 時長 = 原始片段時長。
測試：pytest 29 passed，tsc --noEmit 無錯誤。

+++ backend/app/services/preview.py
@@ -124,7 +124,7 @@ def _track_color(tid: int) -> tuple[int, int, int]:
 def render_overlay_video(
     pkl_path: str | Path,
     output_mp4: str | Path,
-    fps: int = 30,
+    fps: float = 30,
 ) -> Path:
     import cv2
